pyproject.toml
[project]
name = "otel-instrumentation-mcp"
version = "0.15.0"
description = "OpenTelemetry Instrumentation MCP Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.1.0",
    "fastmcp>=0.5.0",
    "fastapi>=0.115.0",
    "opentelemetry-api>=1.35.0",
    "opentelemetry-exporter-otlp>=1.35.0",
    "opentelemetry-sdk>=1.35.0",
    "opentelemetry-instrumentation-requests>=0.46b0",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "opentelemetry-instrumentation-logging>=0.46b0",
    "opentelemetry-semantic-conventions>=0.46b0",
    "uvicorn>=0.35.0",
    "requests>=2.32.0",
    "markdown>=3.7.0",
    "beautifulsoup4>=4.12.0",
    "PyJWT>=2.8.0",
]

[dependency-groups]
dev = [
    "black>=24.0.0",
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]

[project.scripts]
otel-instrumentation-mcp = "otel_instrumentation_mcp.main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["otel_instrumentation_mcp*"]
